// libs/enums/product.enum.ts
export enum ProductType {
  CLOTHING = "CLOTHING",
  HOME_LIVING = "HOME AND LIVING",
  ACCESSORY = "ACCESSORY",
  HANDMADE = "HANDMADE",
  VINTAGE = "VINTAGE",
  CRAFT_SUPPLIES = "CRAFT SUPPLIES",
  JEWELRY = "JEWELRY",
  PET_PRODUCTS = "PET PRODUCTS",
  ART_COLLECTABLES = "ART AND COLLECTABLES",
  CHILDREN = "CHILDREN",
}

export enum ProductStatus {
  HOLD = "HOLD",
  ACTIVE = "ACTIVE",
  SOLD = "SOLD",
  DELETE = "DELETE",
}

export enum SHippingTimeType {
  TURBO = "24 hours",
  FAST = "3-5 days",
  SLOW = "7-10 days",
}